Add a perf_sigmoid_fitter.py script
This script finds the sigmoid scaling that matches the provided data. The sigmoid scaling is the coefficient k in sigmoid(eval*k) such that the resulting value matches the perf% derived from scores and outcomes.

This script can be used as a standalone tool by calling it with a filename as an argument (no arguments to see help); or as a module in which case it can be used to compute the correct sigmoid scaling for a training batch.