Attempt to optimize the position skipping to fit a distribution.
The old algorithm work online and therefore might discard excessive amounts of positions.
The new algorithm filters the buffers that are filled in each worker thread
(currently each buffer is around 1M elements). This offline algorithm allows more exact
distribution match and less wasted positions. It is however tied to the implementation of the buffering,
which is not done for .bin files (so it currently is not implemented for .bin files), and the buffer size will
now have (a small) effect. Similarly FEN stream filtering no longer uses this type of skipping.