Fix Weights Loading Torch
In PyTorch 2.6, we changed the default value of the `weights_only`
argument in `torch.load` from `False` to `True`. Re-running `torch.load`
with `weights_only` set to `False` will likely succeed, but it can
result in arbitrary code execution. Do it only if you got the file from
a trusted source.